Split test cases to tasks
